[IMP] hw_posbox_upgrade: Integrate the hw_posbox_upgrade module in hw_posbox homepage
We move the main.py controller and the upgrade.html view
in the hw_posbox_homepage module for more consistency
